Fix RGBV encoding default and PCG hash type safety
- Change RGBV_DefaultS from 1.f to 0.f for correct encoding default
- Add explicit uint32_t cast in PCGHashRand to ensure type safety

## codocs-skip
- .codocs/docs/include/UCommon/Codec.h.md: 仅修改常数值，不影响编码算法说明与接口文档
- .codocs/docs/include/UCommon/Utils.h.md: 仅改进函数实现的类型安全性，不改变函数语义与输出范围文档

+++ include/UCommon/Utils.h
@@ -280,7 +280,7 @@ namespace UCommon
 
 	static inline float PCGHashRand(uint32_t Input)
 	{
-		return (PCGHash(Input) % RAND_MAX) / (float)RAND_MAX;
+		return (float)(PCGHash(Input) % (uint32_t)RAND_MAX) / (float)RAND_MAX;
 	}
 
 	enum class EElementType : std::uint64_t
